add true cli support
